requirements.txt
pyqt6
pillow
pyautogui
keyboard
pyinstaller
